Resolves #256 google analytics pageviews for file_sets
Much of this work has been lifted from Sufia, especially services/analytics_service.rb,
model/pageview.rb and their respective specs. But not all of the Sufia functionality
is carried over, there's no statistics.rb model for instance. There's no model for stats,
the service is just called directly from the presenter.

It might be that as demands for more complicated analytics functionality come in
we'll want to refactor and look for ways to manage that complexity. When that happens
it probably makes sense to look to Sufia and try to emulate more of what they've done.

Unless we get a plugin archetecture before that happens and can just include a GA gem(!)