Resolves #481: add ga event tracking
This is all done in javascript, similar to the old blacklight_google_analytics gem:

https://github.com/uohull/blacklight_google_analytics/blob/master/app/assets/javascripts/blacklight_google_analytics.js.erb

The event tracking relies on the DOM being stable. There is currently no javascript test framework
in heliotrope, so there are rspec assertions in different feature specs that make sure the required
DOM element exists. If those tests fail, the rspec tests and .js need to be updated with the
changed DOM selectors.