Story #673 Refactored the test setup on some specs to speed up the specs
* For test setup, create the Monograph, Section, and FileSet records in
let blocks instead of creating them by interacting with the UI in the
body of the spec.  (Avoid extra routing and rendering work to speed up
the spec.)

* For specs that don't need admin-level access, don't log in with any
user.  (This change isn't for performance, but just to remove
unnecessary code.  You shouldn't have to be an admin user to view the
facets.)