Resolves #879 Adds fields for primary editor first and last name
We need to differentiate editors from "creators"/authors.

Because of how rdf works, we can't rely on order in the monograph.editor field.

So, like with "creator" we create first and last name fields for the "primary_editor"
and put the first editor in them, with remaining editors in the (unordered)
"editor" list.

This doesn't yet work with the importer, manual editing would need to happen after
importing a new monograph with multiple editors.

Because editors are not used in file_sets, we don't include this in a construct
like the "stores_creator_name_seperately.rb" file. We just put the fields in
the monograph model.

Right now the only use case is a book with 2 editors. When we get to 3, and if
order matters, we'll need to come up with something else...