Resolves #1431 added automated accessibility test infrastructure
Where we originally thought we'd use Deque's Attest software for automated
accessiblity testing, we learned that in order to run it all current
(and crucially future) developers would need credentials from Desque
which could be limiting.

Instead we'll try the open source axe-matchers gem, which relies on the
open source axe-core library (Attest also uses these) and see how far that
gets us. I believe we can't define an individualized ruleset this way,
but we can still use :wcag2a and/or :section508 which will hopefully be fine.

A11y specs will use Rails 5.1 system tests, located in spec/system which in turn
uses headless Chrome and capybara.

A monkey patch is included in this commit: the way heliotrope/hyrax javascript
defines "define" via requireja/almond is not compatible with axe-core's
"axe.js" so a minor change was needed. As we update axe-matchers (and so axe-core)
in the future, we should verify that the monkey patch is still needed.