AEIM-1324 - Add KCV, update Keycard, Checkpoint
This brings in the binding gem, KCV, which has the default
AgentResolver. It also updates the Keycard and Checkpoint gems to
adjust for API changes.

A few of these tests use aggressive mocking with any_instance_of, which
could stand to be cleaned up, but this change just adapts the existing
tests to the updated API. Namely, Keycard::RequestAttributes is no
longer used but a subclass of Keycard::Request::Attributes, the
dlpsInstitutionId key is now a symbol, and #identity is called on the
attributes, rather than #all.