AEIM-1324 - Add Devise / Keycard strategy
- Add new KeycardAuthenticatable strategy, using the user_eid as the
  user_key (alias for email, by way of Blacklight)
- Expose identity as resolved by Keycard on User model
- Remove HttpHeaderAuthenticatable behavior from app controller
- Fix namespacing problem with HttpHeaderAuthenticatable behaviors that
  came up when the controller was no longer autoloading by the wrong
  name before the strategy
- Change valid_user_signed_in? logic to check for the user's email
  which is what it was previously doing, semantically
- Set up FakeAuthHeader to map in REMOTE_USER as well as
  HTTP_X_REMOTE_USER from the dummy header to allow the default
  DirectAttributes to pick it up
- Apply Keycard access mode from Settings

Note that there is a dependency on a Services.request_attributes factory
in the strategy. This is a basic service locator, relying on the
already-registered factory from config/initializers/services.rb. Using
this strategy somewhere else if there was an objection to the Services
constant would probably entail a class variable because the strategy is
instantiated by Warden/Devise and a collaborator is not readily injected
at that point.

The HttpHeaderAuthenticatable strategy is no longer used and should be
extracted and/or removed.

The Authentications controller should be reworked as well. It modifies
ENV on the fly, which does the job for dev, but this type of side effect
is probably better managed with a cookie or some such.