AEIM-1405 - Fix up Shibboleth login and logout
Introduce a flag for "affirmative login", meaning that the user has
actively initiated a login with the application. This signals to the
Devise strategy when it should and should not run.

With Shibboleth protecting the whole application, any time
authenticate_user! ran, if there was an SSO session, the user would
immediately get an application session. This could result in loops where
logout triggered a new session, making it impossible to truly log out.

This change runs the strategy only when the user is actively logging in.
There are lots of edge cases and details here, with comments in the code
and on the issue. This is not the final form, but should be stable. The
login routes, dev login, and error pages should all be cleaned up once
this is confirmed as working.