AEIM-1405 - Embed target in shib login URL
There is apparently something going on with Turbolinks in production,
where there is a double request to shib_login (and then to
Shibboleth.sso/Login). The first one was getting the target correctly,
but the second was failing... So, we go back to embedding the target
into the shib_login URL and shuttle it through the whole process. This
is arguably more consistent, anyway.