HELIO-4120 - Remove therubyracer; depend on node being installed
Everything we do with therubyracer is by way of ExecJS, and NodeJS is picked up
automatically. The libv8 that therubyracer depends on is quite old and there
are some compatibility/compilation problems, in addition to performance
implications. We rely on Yarn and webpack, anyway, so NodeJS is not a new
dependency.